requirements.txt
Django==1.9.5
Markdown==2.6.6
Pillow==3.2.0
Pygments==2.1.3
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
django-filter==0.13.0
django-registration==2.0.4
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
traitlets==4.2.1
